Set up rake tasks to make running features easier
Two things here:

1. I'm going to translate features over, and mark them as failing when
they are failing. I'll also tag them with gh-N to mark GitHub issue #N
that they go along with.
2. 'rake cruise' has been set up to run for CI. Basically, it runs all
the things that aren't failing, and makes sure they pass. Then, it runs
all thet hings that are supposed to be failing, and makes sure they
fail. If either of those are not true, it fails the build, and if
they both pass, they pass. This way we know when something starts
passing again, and to remove the failing tag.