Refactor schema error handling, reintroduce failure
We should handle these errors better, and now that we are,
let's reintroduce the failing second half of the scenario.

I'm actually not entirely sure that flunk-ing with the message
is any different than just bubbling up the error, but at least
it's a real assertion failure error, rather than a JSON API
parsing error.